+++ instrumentation/github.com/aws/aws-sdk-go/service/otels3/s3_test.go
@@ -26,7 +26,6 @@ import (
 	"github.com/aws/aws-sdk-go/service/s3/s3iface"
 	"github.com/stretchr/testify/assert"
 
-	"go.opentelemetry.io/contrib/instrumentation/github.com/aws/aws-sdk-go/service/config"
 	"go.opentelemetry.io/contrib/instrumentation/github.com/aws/aws-sdk-go/service/otels3/mocks"
 	mockmetric "go.opentelemetry.io/contrib/internal/metric"
 	mocktrace "go.opentelemetry.io/contrib/internal/trace"
@@ -60,8 +59,8 @@ func assertMetrics(t *testing.T, mockedMeterImp *mockmetric.MeterImpl) {
 	assert.Equal(t, 2, len(mockedMeterImp.MeasurementBatches))
 
 	metricsFound := map[string]bool{
-		"storage.operation.duration_μs": false,
-		"storage.s3.operation":          false,
+		"aws.s3.operation.duration": false,
+		"aws.s3.operation":          false,
 	}
 	// iterate over metrics to get names
 	for _, measurementBatch := range mockedMeterImp.MeasurementBatches {
@@ -80,9 +79,9 @@ func assertMetrics(t *testing.T, mockedMeterImp *mockmetric.MeterImpl) {
 	}
 }
 
-func assertSpanCorrelationInMetrics(t *testing.T, spanCorrelationInMetrics bool, mockedMeterImp *mockmetric.MeterImpl, span *mocktrace.Span) {
+func assertSpanCorrelation(t *testing.T, spanCorrelation bool, mockedMeterImp *mockmetric.MeterImpl, span *mocktrace.Span) {
 	for _, measurementBatch := range mockedMeterImp.MeasurementBatches {
-		if spanCorrelationInMetrics {
+		if spanCorrelation {
 			traceID := span.SpanContext().TraceID.String()
 			spanID := span.SpanContext().SpanID.String()
 
@@ -97,8 +96,8 @@ func assertSpanCorrelationInMetrics(t *testing.T, spanCorrelationInMetrics bool,
 
 func Test_instrumentedS3_PutObjectWithContext(t *testing.T) {
 	type fields struct {
-		spanCorrelationInMetrics bool
-		mockSetup                func() (expectedReturn interface{})
+		spanCorrelation bool
+		mockSetup       func() (expectedReturn interface{})
 	}
 	type args struct {
 		ctx   aws.Context
@@ -114,7 +113,7 @@ func Test_instrumentedS3_PutObjectWithContext(t *testing.T) {
 		{
 			name: "instrumentedS3.PutObjectWithContext should be delegated to S3.PutObjectWithContext while metrics and spans are linked",
 			fields: fields{
-				spanCorrelationInMetrics: true,
+				spanCorrelation: true,
 				mockSetup: func() (expectedReturn interface{}) {
 					expectedReturn = &s3.PutObjectOutput{}
 					return
@@ -132,7 +131,7 @@ func Test_instrumentedS3_PutObjectWithContext(t *testing.T) {
 		{
 			name: "instrumentedS3.PutObjectWithContext should be delegated to S3.PutObjectWithContext while metrics and spans are NOT linked",
 			fields: fields{
-				spanCorrelationInMetrics: false,
+				spanCorrelation: false,
 				mockSetup: func() (expectedReturn interface{}) {
 					expectedReturn = &s3.PutObjectOutput{}
 					return
@@ -160,13 +159,13 @@ func Test_instrumentedS3_PutObjectWithContext(t *testing.T) {
 
 			s3Mock := &mocks.MockS3Client{}
 			s := &instrumentedS3{
-				S3API:                    s3Mock,
-				tracer:                   mockedTracer,
-				meter:                    mockedMeter,
-				propagators:              mockedPropagators,
-				counters:                 mockedCounters,
-				recorders:                mockedRecorders,
-				spanCorrelationInMetrics: tt.fields.spanCorrelationInMetrics,
+				S3API:           s3Mock,
+				tracer:          mockedTracer,
+				meter:           mockedMeter,
+				propagators:     mockedPropagators,
+				counters:        mockedCounters,
+				recorders:       mockedRecorders,
+				spanCorrelation: tt.fields.spanCorrelation,
 			}
 			expectedReturn := tt.fields.mockSetup()
 			got, err := s.PutObjectWithContext(tt.args.ctx, tt.args.input)
@@ -189,15 +188,15 @@ func Test_instrumentedS3_PutObjectWithContext(t *testing.T) {
 
 			assertMetrics(t, mockedMeterImp)
 
-			assertSpanCorrelationInMetrics(t, tt.fields.spanCorrelationInMetrics, mockedMeterImp, spans[0])
+			assertSpanCorrelation(t, tt.fields.spanCorrelation, mockedMeterImp, spans[0])
 		})
 	}
 }
 
 func Test_instrumentedS3_GetObjectWithContext(t *testing.T) {
 	type fields struct {
-		spanCorrelationInMetrics bool
-		mockSetup                func() (expectedReturn interface{})
+		spanCorrelation bool
+		mockSetup       func() (expectedReturn interface{})
 	}
 	type args struct {
 		ctx   aws.Context
@@ -213,7 +212,7 @@ func Test_instrumentedS3_GetObjectWithContext(t *testing.T) {
 		{
 			name: "instrumentedS3.GetObjectWithContext should be delegated to S3.GetObjectWithContext while metrics and spans are linked",
 			fields: fields{
-				spanCorrelationInMetrics: true,
+				spanCorrelation: true,
 				mockSetup: func() (expectedReturn interface{}) {
 					expectedReturn = &s3.GetObjectOutput{}
 					return
@@ -231,7 +230,7 @@ func Test_instrumentedS3_GetObjectWithContext(t *testing.T) {
 		{
 			name: "instrumentedS3.GetObjectWithContext should be delegated to S3.GetObjectWithContext while metrics and spans are NOT linked",
 			fields: fields{
-				spanCorrelationInMetrics: false,
+				spanCorrelation: false,
 				mockSetup: func() (expectedReturn interface{}) {
 					expectedReturn = &s3.GetObjectOutput{}
 					return
@@ -259,13 +258,13 @@ func Test_instrumentedS3_GetObjectWithContext(t *testing.T) {
 
 			s3Mock := &mocks.MockS3Client{}
 			s := &instrumentedS3{
-				S3API:                    s3Mock,
-				tracer:                   mockedTracer,
-				meter:                    mockedMeter,
-				propagators:              mockedPropagators,
-				counters:                 mockedCounters,
-				recorders:                mockedRecorders,
-				spanCorrelationInMetrics: tt.fields.spanCorrelationInMetrics,
+				S3API:           s3Mock,
+				tracer:          mockedTracer,
+				meter:           mockedMeter,
+				propagators:     mockedPropagators,
+				counters:        mockedCounters,
+				recorders:       mockedRecorders,
+				spanCorrelation: tt.fields.spanCorrelation,
 			}
 			expectedReturn := tt.fields.mockSetup()
 			got, err := s.GetObjectWithContext(tt.args.ctx, tt.args.input)
@@ -288,15 +287,15 @@ func Test_instrumentedS3_GetObjectWithContext(t *testing.T) {
 
 			assertMetrics(t, mockedMeterImp)
 
-			assertSpanCorrelationInMetrics(t, tt.fields.spanCorrelationInMetrics, mockedMeterImp, spans[0])
+			assertSpanCorrelation(t, tt.fields.spanCorrelation, mockedMeterImp, spans[0])
 		})
 	}
 }
 
 func Test_instrumentedS3_DeleteObjectWithContext(t *testing.T) {
 	type fields struct {
-		spanCorrelationInMetrics bool
-		mockSetup                func() (expectedReturn interface{})
+		spanCorrelation bool
+		mockSetup       func() (expectedReturn interface{})
 	}
 	type args struct {
 		ctx   aws.Context
@@ -312,7 +311,7 @@ func Test_instrumentedS3_DeleteObjectWithContext(t *testing.T) {
 		{
 			name: "instrumentedS3.DeleteObjectWithContext should be delegated to S3.DeleteObjectWithContext while metrics and spans are linked",
 			fields: fields{
-				spanCorrelationInMetrics: true,
+				spanCorrelation: true,
 				mockSetup: func() (expectedReturn interface{}) {
 					expectedReturn = &s3.DeleteObjectOutput{}
 					return
@@ -330,7 +329,7 @@ func Test_instrumentedS3_DeleteObjectWithContext(t *testing.T) {
 		{
 			name: "instrumentedS3.DeleteObjectWithContext should be delegated to S3.DeleteObjectWithContext while metrics and spans are NOT linked",
 			fields: fields{
-				spanCorrelationInMetrics: false,
+				spanCorrelation: false,
 				mockSetup: func() (expectedReturn interface{}) {
 					expectedReturn = &s3.DeleteObjectOutput{}
 					return
@@ -358,13 +357,13 @@ func Test_instrumentedS3_DeleteObjectWithContext(t *testing.T) {
 
 			s3Mock := &mocks.MockS3Client{}
 			s := &instrumentedS3{
-				S3API:                    s3Mock,
-				tracer:                   mockedTracer,
-				meter:                    mockedMeter,
-				propagators:              mockedPropagators,
-				counters:                 mockedCounters,
-				recorders:                mockedRecorders,
-				spanCorrelationInMetrics: tt.fields.spanCorrelationInMetrics,
+				S3API:           s3Mock,
+				tracer:          mockedTracer,
+				meter:           mockedMeter,
+				propagators:     mockedPropagators,
+				counters:        mockedCounters,
+				recorders:       mockedRecorders,
+				spanCorrelation: tt.fields.spanCorrelation,
 			}
 			expectedReturn := tt.fields.mockSetup()
 			got, err := s.DeleteObjectWithContext(tt.args.ctx, tt.args.input)
@@ -387,7 +386,7 @@ func Test_instrumentedS3_DeleteObjectWithContext(t *testing.T) {
 
 			assertMetrics(t, mockedMeterImp)
 
-			assertSpanCorrelationInMetrics(t, tt.fields.spanCorrelationInMetrics, mockedMeterImp, spans[0])
+			assertSpanCorrelation(t, tt.fields.spanCorrelation, mockedMeterImp, spans[0])
 		})
 	}
 }
@@ -395,7 +394,7 @@ func Test_instrumentedS3_DeleteObjectWithContext(t *testing.T) {
 func Test_instrumentedS3_NewInstrumentedS3Client(t *testing.T) {
 	type args struct {
 		s    s3iface.S3API
-		opts []config.Option
+		opts []Option
 	}
 	tracerProvider, _ := mocktrace.NewTracerProviderAndTracer(instrumentationName)
 	_, meterProvider := mockmetric.NewMeterProvider()
@@ -411,16 +410,16 @@ func Test_instrumentedS3_NewInstrumentedS3Client(t *testing.T) {
 			name: "NewInstrumentedS3Client should return a new client",
 			args: args{
 				s: s3MockClient,
-				opts: []config.Option{
-					config.WithTracerProvider(tracerProvider),
-					config.WithMeterProvider(meterProvider),
-					config.WithPropagators(mockedPropagator),
-					config.WithSpanCorrelationInMetrics(true),
+				opts: []Option{
+					WithTracerProvider(tracerProvider),
+					WithMeterProvider(meterProvider),
+					WithPropagators(mockedPropagator),
+					WithSpanCorrelation(true),
 				},
 			},
 			verifyFunc: func(t *testing.T, got s3iface.S3API, err error) {
 				assert.Nil(t, err, "error should be nil")
-				assert.Equal(t, got.(*instrumentedS3).spanCorrelationInMetrics, true)
+				assert.Equal(t, got.(*instrumentedS3).spanCorrelation, true)
 				assert.Equal(t, got.(*instrumentedS3).propagators, mockedPropagator)
 				assert.Equal(t, got.(*instrumentedS3).S3API, s3MockClient)
 				assert.NotNil(t, got.(*instrumentedS3).meter, "meter should not be nil")